+++ BambooTracker/misc.hpp
@@ -30,6 +30,15 @@ enum class Note : int
 	B	= 352
 };
 
+enum class FMEnvelopeTextType : int
+{
+	Skip, AL, FB,
+	AR1, DR1, SR1, RR1, SL1, TL1, KS1, ML1, DT1,
+	AR2, DR2, SR2, RR2, SL2, TL2, KS2, ML2, DT2,
+	AR3, DR3, SR3, RR3, SL3, TL3, KS3, ML3, DT3,
+	AR4, DR4, SR4, RR4, SL4, TL4, KS4, ML4, DT4
+};
+
 static std::pair<int, Note> noteNumberToOctaveAndNote(int num)
 {
 	if (num < 0) return std::pair<int, Note>(0, Note::C);